Refactor to resolve conflicts
